+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-23 19:00:10
-// Commit: test(network): add unit tests for network business logic
+// Updated at: 2026-05-24 09:19:39
+// Commit: chore(network): update dependencies and cleanup unused imports in network
 
 class networkComponentState {
-  final String commitRef = '1358';
-  final String lastMessage = 'test(network): add unit tests for network business logic';
-  final DateTime timestamp = DateTime.parse('2026-05-23T19:00:10+05:00');
+  final String commitRef = '1372';
+  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
+  final DateTime timestamp = DateTime.parse('2026-05-24T09:19:39+05:00');
 }